requirement_MZ.txt
opencv-python             
pandas
pillow
scikit-image
scikit-learn
scikit-learn-intelex
scipy 
python-color-transfer
streamlit
pytorch